Camera added

+++ include/Component.h
@@ -9,7 +9,7 @@ class Component{
 public:
     Component(GameObject& associated);
     virtual ~Component();
-    void virtual Update(float dt) = 0;
+    void virtual Update(int dt) = 0;
     void virtual Render() = 0;
     bool virtual Is(std::string type) = 0;
