+++ include/Face.h
@@ -9,7 +9,7 @@ class Face : public Component {
 public:
     Face(GameObject& associated);
     void Damage(int damage);
-    void Update(float dt);
+    void Update(int dt);
     void Render();
     bool Is(std::string type);
